+++ src/edu/univ/erp/ui/admin/manageCourseSection/createCourse.java
@@ -1,4 +1,4 @@
-package edu.univ.erp.ui.admin;
+package edu.univ.erp.ui.admin.manageCourseSection;
 
 import javax.swing.BorderFactory;
 import javax.swing.JButton;
@@ -12,6 +12,7 @@ import javax.swing.UIManager;
 import com.formdev.flatlaf.FlatLightLaf;
 
 import edu.univ.erp.service.AdminService;
+import edu.univ.erp.ui.admin.adminDashboard;
 import edu.univ.erp.util.BREATHEFONT;
 
 import java.awt.BorderLayout;
@@ -24,10 +25,11 @@ import java.awt.Insets;
 import java.awt.event.ActionListener;
 import java.awt.event.ActionEvent;
 
-public class deleteSectionACourse {
-    public deleteSectionACourse(String roll_no) { 
+public class createCourse {
 
-         Font breatheFont = BREATHEFONT.fontGen();
+    public createCourse(String roll_no) {
+
+        Font breatheFont = BREATHEFONT.fontGen();
         Font gFont = BREATHEFONT.gFontGen(); 
 
         try {
@@ -36,8 +38,8 @@ public class deleteSectionACourse {
             e.printStackTrace();
         }
 
-        JFrame f = new JFrame("Delete Section and Course");
-        f.setSize(800, 600); 
+        JFrame f = new JFrame("Create Course");
+        f.setSize(800, 600);
         f.setLayout(new BorderLayout());
 
         // ---- TOP ----
@@ -45,9 +47,9 @@ public class deleteSectionACourse {
         p1.setBackground(Color.decode("#051072")); 
         p1.setOpaque(true); 
         
-        JLabel l0 = new JLabel("DELETE COURSE"); 
+        JLabel l0 = new JLabel("CREATE COURSE"); 
         l0.setForeground(Color.decode("#dbd3c5"));
-        l0.setFont(breatheFont.deriveFont(Font.BOLD, 60f)); 
+        l0.setFont(breatheFont.deriveFont(Font.BOLD, 80f)); 
         l0.setBorder(BorderFactory.createEmptyBorder(15, 0, 0, 0));
         p1.add(l0);
         f.add(p1, BorderLayout.NORTH);
@@ -55,10 +57,10 @@ public class deleteSectionACourse {
         // ---- MIDDLE ----
         JPanel p2 = new JPanel(new GridBagLayout());
         p2.setBackground(Color.decode("#dbd3c5")); 
-        p2.setBorder(BorderFactory.createEmptyBorder(50, 100, 50, 100));
+        p2.setBorder(BorderFactory.createEmptyBorder(20, 50, 20, 50));
 
         GridBagConstraints gbc = new GridBagConstraints();
-        gbc.insets = new Insets(15, 15, 15, 15); 
+        gbc.insets = new Insets(10, 10, 10, 10); 
         gbc.fill = GridBagConstraints.HORIZONTAL;
 
         // Row 0: Course Code
@@ -75,40 +77,67 @@ public class deleteSectionACourse {
         gbc.anchor = GridBagConstraints.WEST;
         p2.add(tCourseCode, gbc);
 
-        // Row 1: Section
-        JLabel l2 = new JLabel("Section:");
+        // Row 1: Title
+        JLabel l2 = new JLabel("Title:");
         l2.setFont(gFont.deriveFont(Font.PLAIN, 24));
         l2.setForeground(Color.decode("#020A48"));
         gbc.gridx = 0; gbc.gridy = 1; gbc.weightx = 0;
         gbc.anchor = GridBagConstraints.EAST;
         p2.add(l2, gbc);
 
-        JTextField tSection = new JTextField(20);
-        tSection.setFont(gFont.deriveFont(Font.PLAIN, 21));
+        JTextField tTitle = new JTextField(20); 
+        tTitle.setFont(gFont.deriveFont(Font.PLAIN, 21));
         gbc.gridx = 1; gbc.gridy = 1; gbc.weightx = 1;
         gbc.anchor = GridBagConstraints.WEST;
+        p2.add(tTitle, gbc);
+
+        // Row 2: Section
+        JLabel l3 = new JLabel("Section:");
+        l3.setFont(gFont.deriveFont(Font.PLAIN, 24));
+        l3.setForeground(Color.decode("#020A48"));
+        gbc.gridx = 0; gbc.gridy = 2; gbc.weightx = 0;
+        gbc.anchor = GridBagConstraints.EAST;
+        p2.add(l3, gbc);
+
+        JTextField tSection = new JTextField(20); 
+        tSection.setFont(gFont.deriveFont(Font.PLAIN, 21));
+        gbc.gridx = 1; gbc.gridy = 2; gbc.weightx = 1;
+        gbc.anchor = GridBagConstraints.WEST;
         p2.add(tSection, gbc);
+
+        // Row 3: Credits
+        JLabel l4 = new JLabel("Credits:");
+        l4.setFont(gFont.deriveFont(Font.PLAIN, 24));
+        l4.setForeground(Color.decode("#020A48"));
+        gbc.gridx = 0; gbc.gridy = 3; gbc.weightx = 0;
+        gbc.anchor = GridBagConstraints.EAST;
+        p2.add(l4, gbc);
+
+        JTextField tCredits = new JTextField(20); 
+        tCredits.setFont(gFont.deriveFont(Font.PLAIN, 21));
+        gbc.gridx = 1; gbc.gridy = 3; gbc.weightx = 1;
+        gbc.anchor = GridBagConstraints.WEST;
+        p2.add(tCredits, gbc);
 
         // --- BUTTONS ---
         JPanel buttonPanel = new JPanel(new FlowLayout(FlowLayout.CENTER, 20, 0));
         buttonPanel.setOpaque(false);
 
-        JButton bDelete = new JButton("Delete");
-        bDelete.setBackground(Color.decode("#051072")); 
-        bDelete.setForeground(Color.WHITE);
-        bDelete.setFont(breatheFont.deriveFont(Font.PLAIN, 35));
-        bDelete.setMargin(new Insets(10, 30, 5, 30));
-        buttonPanel.add(bDelete);
-
-        JButton bBack = new JButton("Back");
-        bBack.setBackground(Color.decode("#2f77b1"));
-        bBack.setForeground(Color.WHITE);
+        JButton bAssign = new JButton("Create"); 
+        bAssign.setBackground(Color.decode("#2f77b1")); 
+        bAssign.setForeground(Color.WHITE); 
+        bAssign.setFont(breatheFont.deriveFont(Font.PLAIN, 35));
+        bAssign.setMargin(new Insets(10, 30, 5, 30));
+        buttonPanel.add(bAssign);
+        
+        JButton bBack = new JButton("Back"); 
+        bBack.setBackground(Color.decode("#2f77b1")); 
+        bBack.setForeground(Color.WHITE); 
         bBack.setFont(breatheFont.deriveFont(Font.PLAIN, 35));
         bBack.setMargin(new Insets(10, 30, 5, 30));
         buttonPanel.add(bBack);
 
-        // Place buttons
-        gbc.gridx = 0; gbc.gridy = 2; 
+        gbc.gridx = 0; gbc.gridy = 4; 
         gbc.gridwidth = 2; 
         gbc.anchor = GridBagConstraints.CENTER;
         gbc.fill = GridBagConstraints.NONE;
@@ -122,45 +151,29 @@ public class deleteSectionACourse {
         f.setVisible(true);
 
         // --- Action Listeners ---
-        bDelete.addActionListener(new ActionListener() {
+        bAssign.addActionListener(new ActionListener() {
             public void actionPerformed(ActionEvent e) {
-                String courseCode = tCourseCode.getText().trim();
-                String section = tSection.getText().trim();
+                String iCode = tCourseCode.getText().trim();
+                String iTitle = tTitle.getText().trim();
+                String iSec = tSection.getText().trim();
+                String iCredits = tCredits.getText().trim();
 
-                if (courseCode.isEmpty() || section.isEmpty()) {
-                    JOptionPane.showMessageDialog(null, "Please enter both Course Code and Section.", "Error", JOptionPane.ERROR_MESSAGE);
-                    return;
-                }
+                AdminService service = new AdminService();
+                String result = service.createCourse(iCode, iTitle, iSec, iCredits);
 
-                int confirm = JOptionPane.showConfirmDialog(null, 
-                    "Are you sure you want to delete " + courseCode + " section " + section + "?\nThis will remove it from both Courses and Sections tables.", 
-                    "Confirm Delete", 
-                    JOptionPane.YES_NO_OPTION);
-
-                if (confirm == JOptionPane.YES_OPTION) {
-                    // 1. Call Service
-                    AdminService service = new AdminService();
-                    String message = service.deleteCourseAndSection(courseCode, section);
-                    
-                    // 2. Show Result
-                    if (message.startsWith("Success")) {
-                        JOptionPane.showMessageDialog(null, message, "Success", JOptionPane.INFORMATION_MESSAGE);
-                        new adminDashboard(roll_no);
-                        f.dispose();
-                    } else if (message.startsWith("Warning")) {
-                        JOptionPane.showMessageDialog(null, message, "Warning", JOptionPane.WARNING_MESSAGE);
-                        new adminDashboard(roll_no);
-                        f.dispose();
-                    } else {
-                        JOptionPane.showMessageDialog(null, message, "Error", JOptionPane.ERROR_MESSAGE);
-                    }
+                if ("Success".equals(result)) {
+                    JOptionPane.showMessageDialog(null, "Course created successfully!", "Success", JOptionPane.INFORMATION_MESSAGE);
+                    new adminDashboard(roll_no);
+                    f.dispose();
+                } else {
+                    JOptionPane.showMessageDialog(null, result, "Error", JOptionPane.ERROR_MESSAGE);
                 }
             }
         });
 
         bBack.addActionListener(new ActionListener() {
             public void actionPerformed(ActionEvent e) {
-                System.out.println("Going back...");
+                System.out.println("Going back to Admin Dashboard..");
                 new adminDashboard(roll_no);
                 f.dispose();
             }
