+++ src/edu/univ/erp/ui/admin/manageCourseSection/deleteSectionACourse.java
@@ -1,4 +1,4 @@
-package edu.univ.erp.ui.admin;
+package edu.univ.erp.ui.admin.manageCourseSection;
 
 import javax.swing.BorderFactory;
 import javax.swing.JButton;
@@ -12,6 +12,7 @@ import javax.swing.UIManager;
 import com.formdev.flatlaf.FlatLightLaf;
 
 import edu.univ.erp.service.AdminService;
+import edu.univ.erp.ui.admin.adminDashboard;
 import edu.univ.erp.util.BREATHEFONT;
 
 import java.awt.BorderLayout;
@@ -24,11 +25,10 @@ import java.awt.Insets;
 import java.awt.event.ActionListener;
 import java.awt.event.ActionEvent;
 
-public class createCourse {
+public class deleteSectionACourse {
+    public deleteSectionACourse(String roll_no) { 
 
-    public createCourse(String roll_no) {
-
-        Font breatheFont = BREATHEFONT.fontGen();
+         Font breatheFont = BREATHEFONT.fontGen();
         Font gFont = BREATHEFONT.gFontGen(); 
 
         try {
@@ -37,8 +37,8 @@ public class createCourse {
             e.printStackTrace();
         }
 
-        JFrame f = new JFrame("Create Course");
-        f.setSize(800, 600);
+        JFrame f = new JFrame("Delete Section and Course");
+        f.setSize(800, 600); 
         f.setLayout(new BorderLayout());
 
         // ---- TOP ----
@@ -46,9 +46,9 @@ public class createCourse {
         p1.setBackground(Color.decode("#051072")); 
         p1.setOpaque(true); 
         
-        JLabel l0 = new JLabel("CREATE COURSE"); 
+        JLabel l0 = new JLabel("DELETE COURSE"); 
         l0.setForeground(Color.decode("#dbd3c5"));
-        l0.setFont(breatheFont.deriveFont(Font.BOLD, 80f)); 
+        l0.setFont(breatheFont.deriveFont(Font.BOLD, 60f)); 
         l0.setBorder(BorderFactory.createEmptyBorder(15, 0, 0, 0));
         p1.add(l0);
         f.add(p1, BorderLayout.NORTH);
@@ -56,10 +56,10 @@ public class createCourse {
         // ---- MIDDLE ----
         JPanel p2 = new JPanel(new GridBagLayout());
         p2.setBackground(Color.decode("#dbd3c5")); 
-        p2.setBorder(BorderFactory.createEmptyBorder(20, 50, 20, 50));
+        p2.setBorder(BorderFactory.createEmptyBorder(50, 100, 50, 100));
 
         GridBagConstraints gbc = new GridBagConstraints();
-        gbc.insets = new Insets(10, 10, 10, 10); 
+        gbc.insets = new Insets(15, 15, 15, 15); 
         gbc.fill = GridBagConstraints.HORIZONTAL;
 
         // Row 0: Course Code
@@ -76,67 +76,40 @@ public class createCourse {
         gbc.anchor = GridBagConstraints.WEST;
         p2.add(tCourseCode, gbc);
 
-        // Row 1: Title
-        JLabel l2 = new JLabel("Title:");
+        // Row 1: Section
+        JLabel l2 = new JLabel("Section:");
         l2.setFont(gFont.deriveFont(Font.PLAIN, 24));
         l2.setForeground(Color.decode("#020A48"));
         gbc.gridx = 0; gbc.gridy = 1; gbc.weightx = 0;
         gbc.anchor = GridBagConstraints.EAST;
         p2.add(l2, gbc);
 
-        JTextField tTitle = new JTextField(20); 
-        tTitle.setFont(gFont.deriveFont(Font.PLAIN, 21));
+        JTextField tSection = new JTextField(20);
+        tSection.setFont(gFont.deriveFont(Font.PLAIN, 21));
         gbc.gridx = 1; gbc.gridy = 1; gbc.weightx = 1;
         gbc.anchor = GridBagConstraints.WEST;
-        p2.add(tTitle, gbc);
-
-        // Row 2: Section
-        JLabel l3 = new JLabel("Section:");
-        l3.setFont(gFont.deriveFont(Font.PLAIN, 24));
-        l3.setForeground(Color.decode("#020A48"));
-        gbc.gridx = 0; gbc.gridy = 2; gbc.weightx = 0;
-        gbc.anchor = GridBagConstraints.EAST;
-        p2.add(l3, gbc);
-
-        JTextField tSection = new JTextField(20); 
-        tSection.setFont(gFont.deriveFont(Font.PLAIN, 21));
-        gbc.gridx = 1; gbc.gridy = 2; gbc.weightx = 1;
-        gbc.anchor = GridBagConstraints.WEST;
         p2.add(tSection, gbc);
-
-        // Row 3: Credits
-        JLabel l4 = new JLabel("Credits:");
-        l4.setFont(gFont.deriveFont(Font.PLAIN, 24));
-        l4.setForeground(Color.decode("#020A48"));
-        gbc.gridx = 0; gbc.gridy = 3; gbc.weightx = 0;
-        gbc.anchor = GridBagConstraints.EAST;
-        p2.add(l4, gbc);
-
-        JTextField tCredits = new JTextField(20); 
-        tCredits.setFont(gFont.deriveFont(Font.PLAIN, 21));
-        gbc.gridx = 1; gbc.gridy = 3; gbc.weightx = 1;
-        gbc.anchor = GridBagConstraints.WEST;
-        p2.add(tCredits, gbc);
 
         // --- BUTTONS ---
         JPanel buttonPanel = new JPanel(new FlowLayout(FlowLayout.CENTER, 20, 0));
         buttonPanel.setOpaque(false);
 
-        JButton bAssign = new JButton("Create"); 
-        bAssign.setBackground(Color.decode("#2f77b1")); 
-        bAssign.setForeground(Color.WHITE); 
-        bAssign.setFont(breatheFont.deriveFont(Font.PLAIN, 35));
-        bAssign.setMargin(new Insets(10, 30, 5, 30));
-        buttonPanel.add(bAssign);
-        
-        JButton bBack = new JButton("Back"); 
-        bBack.setBackground(Color.decode("#2f77b1")); 
-        bBack.setForeground(Color.WHITE); 
+        JButton bDelete = new JButton("Delete");
+        bDelete.setBackground(Color.decode("#051072")); 
+        bDelete.setForeground(Color.WHITE);
+        bDelete.setFont(breatheFont.deriveFont(Font.PLAIN, 35));
+        bDelete.setMargin(new Insets(10, 30, 5, 30));
+        buttonPanel.add(bDelete);
+
+        JButton bBack = new JButton("Back");
+        bBack.setBackground(Color.decode("#2f77b1"));
+        bBack.setForeground(Color.WHITE);
         bBack.setFont(breatheFont.deriveFont(Font.PLAIN, 35));
         bBack.setMargin(new Insets(10, 30, 5, 30));
         buttonPanel.add(bBack);
 
-        gbc.gridx = 0; gbc.gridy = 4; 
+        // Place buttons
+        gbc.gridx = 0; gbc.gridy = 2; 
         gbc.gridwidth = 2; 
         gbc.anchor = GridBagConstraints.CENTER;
         gbc.fill = GridBagConstraints.NONE;
@@ -150,29 +123,38 @@ public class createCourse {
         f.setVisible(true);
 
         // --- Action Listeners ---
-        bAssign.addActionListener(new ActionListener() {
+        bDelete.addActionListener(new ActionListener() {
             public void actionPerformed(ActionEvent e) {
-                String iCode = tCourseCode.getText().trim();
-                String iTitle = tTitle.getText().trim();
-                String iSec = tSection.getText().trim();
-                String iCredits = tCredits.getText().trim();
+                String courseCode = tCourseCode.getText().trim();
+                String section = tSection.getText().trim();
 
+                if (courseCode.isEmpty() || section.isEmpty()) {
+                    JOptionPane.showMessageDialog(null, "Please enter both Course Code and Section.", "Error", JOptionPane.ERROR_MESSAGE);
+                    return;
+                }
+
+                // 1. Call Service
                 AdminService service = new AdminService();
-                String result = service.createCourse(iCode, iTitle, iSec, iCredits);
-
-                if ("Success".equals(result)) {
-                    JOptionPane.showMessageDialog(null, "Course created successfully!", "Success", JOptionPane.INFORMATION_MESSAGE);
+                String message = service.deleteCourseAndSection(courseCode, section);
+                
+                // 2. Show Result
+                if (message.startsWith("Success")) {
+                    JOptionPane.showMessageDialog(null, message, "Success", JOptionPane.INFORMATION_MESSAGE);
+                    new adminDashboard(roll_no);
+                    f.dispose();
+                } else if (message.startsWith("Warning")) {
+                    JOptionPane.showMessageDialog(null, message, "Warning", JOptionPane.WARNING_MESSAGE);
                     new adminDashboard(roll_no);
                     f.dispose();
                 } else {
-                    JOptionPane.showMessageDialog(null, result, "Error", JOptionPane.ERROR_MESSAGE);
+                    JOptionPane.showMessageDialog(null, message, "Error", JOptionPane.ERROR_MESSAGE);
                 }
             }
         });
 
         bBack.addActionListener(new ActionListener() {
             public void actionPerformed(ActionEvent e) {
-                System.out.println("Going back to Admin Dashboard..");
+                System.out.println("Going back...");
                 new adminDashboard(roll_no);
                 f.dispose();
             }
