+++ src/edu/univ/erp/ui/admin/manageCourseSection/editCourse.java
@@ -1,4 +1,4 @@
-package edu.univ.erp.ui.admin;
+package edu.univ.erp.ui.admin.manageCourseSection;
 
 import javax.swing.BorderFactory;
 import javax.swing.JButton;
@@ -11,24 +11,25 @@ import javax.swing.UIManager;
 
 import com.formdev.flatlaf.FlatLightLaf;
 
-import edu.univ.erp.domain.Sections;
+import edu.univ.erp.domain.Course;
 import edu.univ.erp.service.AdminService;
+import edu.univ.erp.ui.admin.adminDashboard;
 import edu.univ.erp.util.BREATHEFONT;
 
-import java.awt.Color;
-import java.awt.Font;
-import java.awt.event.ActionListener;
-import java.sql.SQLException;
-import java.awt.event.ActionEvent;
 import java.awt.BorderLayout;
+import java.awt.Color;
 import java.awt.FlowLayout;
+import java.awt.Font;
 import java.awt.GridBagConstraints;
 import java.awt.GridBagLayout;
 import java.awt.Insets;
+import java.awt.event.ActionListener;
+import java.sql.SQLException;
+import java.awt.event.ActionEvent;
 
-public class editSection {
-    public editSection(String roll_no) {
+public class editCourse {
 
+    public editCourse(String roll_no) {
         Font breatheFont = BREATHEFONT.fontGen();
         Font gFont = BREATHEFONT.gFontGen(); 
 
@@ -38,7 +39,7 @@ public class editSection {
             e.printStackTrace();
         }
 
-        JFrame f0 = new JFrame("Edit Section");
+        JFrame f0 = new JFrame("Edit Course");
         f0.setSize(800, 600); 
         f0.setLayout(new BorderLayout());
 
@@ -47,7 +48,7 @@ public class editSection {
         p1.setBackground(Color.decode("#051072")); 
         p1.setOpaque(true); 
         
-        JLabel l0 = new JLabel("EDIT SECTION"); 
+        JLabel l0 = new JLabel("EDIT COURSE"); 
         l0.setForeground(Color.decode("#dbd3c5"));
         l0.setFont(breatheFont.deriveFont(Font.BOLD, 60f)); 
         l0.setBorder(BorderFactory.createEmptyBorder(15, 0, 0, 0));
@@ -102,14 +103,13 @@ public class editSection {
         bContinue.setMargin(new Insets(10, 30, 5, 30));
         buttonPanel.add(bContinue);
 
-        JButton bBack = new JButton("Back");
-        bBack.setBackground(Color.decode("#2f77b1")); 
-        bBack.setForeground(Color.WHITE); 
-        bBack.setFont(breatheFont.deriveFont(Font.PLAIN, 35));
-        bBack.setMargin(new Insets(10, 30, 5, 30));
-        buttonPanel.add(bBack);
+        JButton backButton = new JButton("Back");
+        backButton.setBackground(Color.decode("#2f77b1")); 
+        backButton.setForeground(Color.WHITE); 
+        backButton.setFont(breatheFont.deriveFont(Font.PLAIN, 35));
+        backButton.setMargin(new Insets(10, 30, 5, 30));
+        buttonPanel.add(backButton);
 
-        // Place buttons
         gbc.gridx = 0; gbc.gridy = 2; 
         gbc.gridwidth = 2; 
         gbc.anchor = GridBagConstraints.CENTER;
@@ -122,36 +122,31 @@ public class editSection {
         f0.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
         f0.setLocationRelativeTo(null);
         f0.setVisible(true);
-
+        
         // --- Action Listeners ---
         bContinue.addActionListener(new ActionListener() {
             public void actionPerformed(ActionEvent e) {
-                String courseCode = tCourseCode.getText().trim();
-                String section = tSection.getText().trim();
-
-                if (courseCode.isEmpty() || section.isEmpty()) {
+                String code = tCourseCode.getText().trim();
+                String sec = tSection.getText().trim();
+                if (code.isEmpty() || sec.isEmpty()) {
                     JOptionPane.showMessageDialog(null, "Please enter both Course Code and Section.", "Missing Info", JOptionPane.WARNING_MESSAGE);
                     return;
                 }
+
                 try {
                     AdminService service = new AdminService();
-                    Sections s = service.getSectionDetails(courseCode, section);
-                    
-                    if (s != null) {
-                        String[] arr = new String[8];
-                        arr[0] = s.getCourseCode();
-                        arr[1] = s.getSection();
-                        arr[2] = s.getRollNo();
-                        arr[3] = s.getDayTime();
-                        arr[4] = s.getRoom();
-                        arr[5] = s.getCapacity();
-                        arr[6] = s.getSemester();
-                        arr[7] = s.getYear();
+                    Course courseObj = service.getCourseDetails(code, sec);
+                    if (courseObj != null) {
+                        String[] arr = new String[4];
+                        arr[0] = courseObj.getCourseCode();
+                        arr[1] = courseObj.getTitle();
+                        arr[2] = courseObj.getSection();
+                        arr[3] = String.valueOf(courseObj.getCredits());
                         
-                        new editSectionPage2(roll_no, arr);
+                        new editCoursePage2(roll_no, arr);
                         f0.dispose();             
                     } else {
-                        JOptionPane.showMessageDialog(null, "Section not found.", "Error", JOptionPane.ERROR_MESSAGE);
+                        JOptionPane.showMessageDialog(null, "Course not found.", "Error", JOptionPane.ERROR_MESSAGE);
                     }
                 } catch (SQLException ex) {
                     ex.printStackTrace();
@@ -160,12 +155,12 @@ public class editSection {
             }
         });
 
-        bBack.addActionListener(new ActionListener() {
+        backButton.addActionListener(new ActionListener() {
             public void actionPerformed(ActionEvent e) {
                 System.out.println("Going back to Admin Dashboard..");
                 new adminDashboard(roll_no);
                 f0.dispose();
             }
-        });        
+        });                
     }
 }
