+++ src/edu/univ/erp/ui/admin/manageCourseSection/editSection.java
@@ -1,4 +1,4 @@
-package edu.univ.erp.ui.admin;
+package edu.univ.erp.ui.admin.manageCourseSection;
 
 import javax.swing.BorderFactory;
 import javax.swing.JButton;
@@ -11,23 +11,24 @@ import javax.swing.UIManager;
 
 import com.formdev.flatlaf.FlatLightLaf;
 
+import edu.univ.erp.domain.Sections;
 import edu.univ.erp.service.AdminService;
+import edu.univ.erp.ui.admin.adminDashboard;
 import edu.univ.erp.util.BREATHEFONT;
 
 import java.awt.Color;
 import java.awt.Font;
 import java.awt.event.ActionListener;
+import java.sql.SQLException;
 import java.awt.event.ActionEvent;
-
 import java.awt.BorderLayout;
 import java.awt.FlowLayout;
 import java.awt.GridBagConstraints;
 import java.awt.GridBagLayout;
 import java.awt.Insets;
 
-public class editCoursePage2 {
-
-    public editCoursePage2(String roll_no,String arrgs[]) {
+public class editSection {
+    public editSection(String roll_no) {
 
         Font breatheFont = BREATHEFONT.fontGen();
         Font gFont = BREATHEFONT.gFontGen(); 
@@ -38,32 +39,32 @@ public class editCoursePage2 {
             e.printStackTrace();
         }
 
-        JFrame f = new JFrame("Edit Course Details");
-        f.setSize(800, 600); 
-        f.setLayout(new BorderLayout());
+        JFrame f0 = new JFrame("Edit Section");
+        f0.setSize(800, 600); 
+        f0.setLayout(new BorderLayout());
 
         // ---- TOP ----
         JPanel p1 = new JPanel();
         p1.setBackground(Color.decode("#051072")); 
         p1.setOpaque(true); 
         
-        JLabel l0 = new JLabel("EDIT COURSE"); 
+        JLabel l0 = new JLabel("EDIT SECTION"); 
         l0.setForeground(Color.decode("#dbd3c5"));
-        l0.setFont(breatheFont.deriveFont(Font.BOLD, 80f)); 
+        l0.setFont(breatheFont.deriveFont(Font.BOLD, 60f)); 
         l0.setBorder(BorderFactory.createEmptyBorder(15, 0, 0, 0));
         p1.add(l0);
-        f.add(p1, BorderLayout.NORTH);
+        f0.add(p1, BorderLayout.NORTH);
 
         // ---- MIDDLE ----
         JPanel p2 = new JPanel(new GridBagLayout());
         p2.setBackground(Color.decode("#dbd3c5")); 
-        p2.setBorder(BorderFactory.createEmptyBorder(20, 50, 20, 50));
+        p2.setBorder(BorderFactory.createEmptyBorder(50, 100, 50, 100));
 
         GridBagConstraints gbc = new GridBagConstraints();
-        gbc.insets = new Insets(10, 10, 10, 10); 
+        gbc.insets = new Insets(15, 15, 15, 15); 
         gbc.fill = GridBagConstraints.HORIZONTAL;
 
-        // Row 0: Course Code (Non-Editable)
+        // Row 0: Course Code
         JLabel l1 = new JLabel("Course Code:");
         l1.setFont(gFont.deriveFont(Font.PLAIN, 24));
         l1.setForeground(Color.decode("#020A48"));
@@ -73,13 +74,11 @@ public class editCoursePage2 {
 
         JTextField tCourseCode = new JTextField(20);
         tCourseCode.setFont(gFont.deriveFont(Font.PLAIN, 21));
-        tCourseCode.setText(arrgs[0]); // Pre-fill
-        tCourseCode.setEditable(false); // Read-only
         gbc.gridx = 1; gbc.gridy = 0; gbc.weightx = 1;
         gbc.anchor = GridBagConstraints.WEST;
         p2.add(tCourseCode, gbc);
 
-        // Row 1: Section (Non-Editable)
+        // Row 1: Section
         JLabel l2 = new JLabel("Section:");
         l2.setFont(gFont.deriveFont(Font.PLAIN, 24));
         l2.setForeground(Color.decode("#020A48"));
@@ -89,52 +88,20 @@ public class editCoursePage2 {
 
         JTextField tSection = new JTextField(20);
         tSection.setFont(gFont.deriveFont(Font.PLAIN, 21));
-        tSection.setText(arrgs[2]); // Pre-fill
-        tSection.setEditable(false); // Read-only
         gbc.gridx = 1; gbc.gridy = 1; gbc.weightx = 1;
         gbc.anchor = GridBagConstraints.WEST;
         p2.add(tSection, gbc);
-
-        // Row 2: Title (Editable)
-        JLabel l3 = new JLabel("Title:");
-        l3.setFont(gFont.deriveFont(Font.PLAIN, 24));
-        l3.setForeground(Color.decode("#020A48"));
-        gbc.gridx = 0; gbc.gridy = 2; gbc.weightx = 0;
-        gbc.anchor = GridBagConstraints.EAST;
-        p2.add(l3, gbc);
-
-        JTextField tTitle = new JTextField(20);
-        tTitle.setFont(gFont.deriveFont(Font.PLAIN, 21));
-        tTitle.setText(arrgs[1]); // Pre-fill
-        gbc.gridx = 1; gbc.gridy = 2; gbc.weightx = 1;
-        gbc.anchor = GridBagConstraints.WEST;
-        p2.add(tTitle, gbc);
-
-        // Row 3: Credits (Editable)
-        JLabel l4 = new JLabel("Credits:");
-        l4.setFont(gFont.deriveFont(Font.PLAIN, 24));
-        l4.setForeground(Color.decode("#020A48"));
-        gbc.gridx = 0; gbc.gridy = 3; gbc.weightx = 0;
-        gbc.anchor = GridBagConstraints.EAST;
-        p2.add(l4, gbc);
-
-        JTextField tCredits = new JTextField(20);
-        tCredits.setFont(gFont.deriveFont(Font.PLAIN, 21));
-        tCredits.setText(arrgs[3]); // Pre-fill
-        gbc.gridx = 1; gbc.gridy = 3; gbc.weightx = 1;
-        gbc.anchor = GridBagConstraints.WEST;
-        p2.add(tCredits, gbc);
 
         // --- BUTTONS ---
         JPanel buttonPanel = new JPanel(new FlowLayout(FlowLayout.CENTER, 20, 0));
         buttonPanel.setOpaque(false);
 
-        JButton bAssign = new JButton("Save"); 
-        bAssign.setBackground(Color.decode("#2f77b1")); 
-        bAssign.setForeground(Color.WHITE); 
-        bAssign.setFont(breatheFont.deriveFont(Font.PLAIN, 35));
-        bAssign.setMargin(new Insets(10, 30, 5, 30));
-        buttonPanel.add(bAssign);
+        JButton bContinue = new JButton("Continue");
+        bContinue.setBackground(Color.decode("#2f77b1")); 
+        bContinue.setForeground(Color.WHITE); 
+        bContinue.setFont(breatheFont.deriveFont(Font.PLAIN, 35));
+        bContinue.setMargin(new Insets(10, 30, 5, 30));
+        buttonPanel.add(bContinue);
 
         JButton bBack = new JButton("Back");
         bBack.setBackground(Color.decode("#2f77b1")); 
@@ -144,39 +111,52 @@ public class editCoursePage2 {
         buttonPanel.add(bBack);
 
         // Place buttons
-        gbc.gridx = 0; gbc.gridy = 4; 
+        gbc.gridx = 0; gbc.gridy = 2; 
         gbc.gridwidth = 2; 
         gbc.anchor = GridBagConstraints.CENTER;
         gbc.fill = GridBagConstraints.NONE;
         gbc.insets = new Insets(50, 10, 10, 10); 
         p2.add(buttonPanel, gbc);
 
-        f.add(p2, BorderLayout.CENTER);
+        f0.add(p2, BorderLayout.CENTER);
 
-        f.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-        f.setLocationRelativeTo(null);
-        f.setVisible(true);
+        f0.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+        f0.setLocationRelativeTo(null);
+        f0.setVisible(true);
 
         // --- Action Listeners ---
-        bAssign.addActionListener(new ActionListener() {
+        bContinue.addActionListener(new ActionListener() {
             public void actionPerformed(ActionEvent e) {
-                // 1. Gather Data
-                String iTitle = tTitle.getText().trim();
-                String iCredits = tCredits.getText().trim();
-                String iCode = tCourseCode.getText().trim();
-                String iSec = tSection.getText().trim();
+                String courseCode = tCourseCode.getText().trim();
+                String section = tSection.getText().trim();
 
-                // 2. Call Service
-                AdminService service = new AdminService();
-                String result = service.updateCourse(iTitle, iCredits, iCode, iSec);
-
-                // 3. Handle Result
-                if ("Success".equals(result)) {
-                    JOptionPane.showMessageDialog(null, "Course updated successfully", "Success", JOptionPane.INFORMATION_MESSAGE);
-                    new adminDashboard(roll_no);
-                    f.dispose();
-                } else {
-                    JOptionPane.showMessageDialog(null, result, "Error", JOptionPane.ERROR_MESSAGE);
+                if (courseCode.isEmpty() || section.isEmpty()) {
+                    JOptionPane.showMessageDialog(null, "Please enter both Course Code and Section.", "Missing Info", JOptionPane.WARNING_MESSAGE);
+                    return;
+                }
+                try {
+                    AdminService service = new AdminService();
+                    Sections s = service.getSectionDetails(courseCode, section);
+                    
+                    if (s != null) {
+                        String[] arr = new String[8];
+                        arr[0] = s.getCourseCode();
+                        arr[1] = s.getSection();
+                        arr[2] = s.getRollNo();
+                        arr[3] = s.getDayTime();
+                        arr[4] = s.getRoom();
+                        arr[5] = s.getCapacity();
+                        arr[6] = s.getSemester();
+                        arr[7] = s.getYear();
+                        
+                        new editSectionPage2(roll_no, arr);
+                        f0.dispose();             
+                    } else {
+                        JOptionPane.showMessageDialog(null, "Section not found.", "Error", JOptionPane.ERROR_MESSAGE);
+                    }
+                } catch (SQLException ex) {
+                    ex.printStackTrace();
+                    JOptionPane.showMessageDialog(null, "Database Error: " + ex.getMessage(), "Error", JOptionPane.ERROR_MESSAGE);
                 }
             }
         });
@@ -185,8 +165,8 @@ public class editCoursePage2 {
             public void actionPerformed(ActionEvent e) {
                 System.out.println("Going back to Admin Dashboard..");
                 new adminDashboard(roll_no);
-                f.dispose();
+                f0.dispose();
             }
-        });
+        });        
     }
 }
