+++ src/edu/univ/erp/ui/admin/manageUsers/addIns.java
@@ -1,7 +1,8 @@
-package edu.univ.erp.ui.admin;
+package edu.univ.erp.ui.admin.manageUsers;
 
 import edu.univ.erp.auth.HashGenerator;
 import edu.univ.erp.service.AdminService;
+import edu.univ.erp.ui.admin.adminDashboard;
 
 import java.awt.event.ActionListener;
 import java.awt.BorderLayout;
@@ -26,8 +27,9 @@ import com.formdev.flatlaf.FlatLightLaf;
 
 import edu.univ.erp.util.BREATHEFONT;
 
-public class addAdm {
-    public addAdm(String rollNo) {
+public class addIns {
+    public addIns(String rollNo) {
+
         Font breatheFont = BREATHEFONT.fontGen();
         Font gFont = BREATHEFONT.gFontGen(); 
 
@@ -46,7 +48,7 @@ public class addAdm {
         p1.setBackground(Color.decode("#051072")); 
         p1.setOpaque(true); 
         
-        JLabel l0 = new JLabel("ADD ADMIN"); 
+        JLabel l0 = new JLabel("ADD INSTRUCTOR"); 
         l0.setForeground(Color.decode("#dbd3c5"));
         l0.setFont(breatheFont.deriveFont(Font.BOLD, 80f)); 
         l0.setBorder(BorderFactory.createEmptyBorder(15, 0, 0, 0));
@@ -88,10 +90,27 @@ public class addAdm {
         gbc.anchor = GridBagConstraints.WEST;
         p2.add(tPassword, gbc);
 
+        JLabel l3 = new JLabel("Department:");
+        l3.setFont(gFont.deriveFont(Font.PLAIN, 24));
+        l3.setForeground(Color.decode("#020A48"));
+        gbc.gridx = 0; gbc.gridy = 2; gbc.weightx = 0;
+        gbc.anchor = GridBagConstraints.EAST;
+        p2.add(l3, gbc);
+
+        JTextField tDepartment = new JTextField(20);
+        tDepartment.setFont(gFont.deriveFont(Font.PLAIN, 21));
+        gbc.gridx = 1; gbc.gridy = 2; gbc.weightx = 1;
+        gbc.anchor = GridBagConstraints.WEST;
+        p2.add(tDepartment, gbc);
+
+        // --- Empty Row 3 for spacing (to push buttons down) ---
+        // This is done implicitly by adjusting the button panel inset, but adding a spacer helps center the form better
+        
+        // --- Row 4: Buttons ---
         JPanel buttonPanel = new JPanel(new FlowLayout(FlowLayout.CENTER, 20, 0));
         buttonPanel.setOpaque(false);
-        
-        // ---- LOWS ----
+ 
+
         JButton b1 = new JButton("Back");
         b1.setBackground(Color.decode("#2f77b1")); 
         b1.setForeground(Color.WHITE); 
@@ -106,7 +125,8 @@ public class addAdm {
         b2.setMargin(new Insets(10, 30, 5, 30));
         buttonPanel.add(b2);
 
-        gbc.gridx = 0; gbc.gridy = 2; 
+        // Added to row 3 (since we removed row 3/4 from addStd)
+        gbc.gridx = 0; gbc.gridy = 3; 
         gbc.gridwidth = 2; 
         gbc.anchor = GridBagConstraints.CENTER;
         gbc.fill = GridBagConstraints.NONE;
@@ -122,16 +142,18 @@ public class addAdm {
         // ---- Action Listeners ----
         b2.addActionListener(new ActionListener() {
             public void actionPerformed(ActionEvent e) {
-                if (tUsername.getText().isEmpty() || tPassword.getText().trim().isEmpty()) {
+                if (tUsername.getText().isEmpty() || tPassword.getText().trim().isEmpty() || tDepartment.getText().trim().isEmpty()) {
                     JOptionPane.showMessageDialog(null, "All fields must be filled out.", "Error", JOptionPane.ERROR_MESSAGE);
                     return;
                 }
+
                 String username = tUsername.getText();
-                String hashPass = HashGenerator.makeHash(tPassword.getText().trim());
+                String department = tDepartment.getText().trim(); 
+                String hash_pass = HashGenerator.makeHash(tPassword.getText().trim());
 
-                AdminService service = new AdminService();  
-                String result = service.registerAdmin(username, hashPass);
-
+                AdminService service = new AdminService();
+                String result = service.registerInstructor(username, hash_pass, department);
+                
                 if (result.startsWith("Success")) {
                     JOptionPane.showMessageDialog(null, result, "Success", JOptionPane.INFORMATION_MESSAGE);
                     f.dispose();
