+++ src/edu/univ/erp/ui/admin/manageUsers/addStd.java
@@ -1,9 +1,14 @@
-package edu.univ.erp.ui.admin;
+package edu.univ.erp.ui.admin.manageUsers;
 
+import javax.swing.*;
+
+import com.formdev.flatlaf.FlatLightLaf;
+
+import edu.univ.erp.util.BREATHEFONT;
 import edu.univ.erp.auth.HashGenerator;
 import edu.univ.erp.service.AdminService;
+import edu.univ.erp.ui.admin.adminDashboard;
 
-import java.awt.event.ActionListener;
 import java.awt.BorderLayout;
 import java.awt.Color;
 import java.awt.FlowLayout;
@@ -12,25 +17,13 @@ import java.awt.GridBagConstraints;
 import java.awt.GridBagLayout;
 import java.awt.Insets;
 import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
 
-import javax.swing.BorderFactory;
-import javax.swing.JButton;
-import javax.swing.JFrame;
-import javax.swing.JLabel;
-import javax.swing.JOptionPane;
-import javax.swing.JPanel;
-import javax.swing.JTextField;
-import javax.swing.UIManager;
-
-import com.formdev.flatlaf.FlatLightLaf;
-
-import edu.univ.erp.util.BREATHEFONT;
-
-public class addIns {
-    public addIns(String rollNo) {
+public class addStd {
+    public addStd(String rollNo) {
 
         Font breatheFont = BREATHEFONT.fontGen();
-        Font gFont = BREATHEFONT.gFontGen(); 
+        Font gFont = BREATHEFONT.gFontGen();
 
         try {
             UIManager.setLookAndFeel(new FlatLightLaf());
@@ -47,7 +40,7 @@ public class addIns {
         p1.setBackground(Color.decode("#051072")); 
         p1.setOpaque(true); 
         
-        JLabel l0 = new JLabel("ADD INSTRUCTOR"); 
+        JLabel l0 = new JLabel("ADD STUDENT");
         l0.setForeground(Color.decode("#dbd3c5"));
         l0.setFont(breatheFont.deriveFont(Font.BOLD, 80f)); 
         l0.setBorder(BorderFactory.createEmptyBorder(15, 0, 0, 0));
@@ -60,7 +53,7 @@ public class addIns {
         p2.setBorder(BorderFactory.createEmptyBorder(20, 50, 20, 50));
 
         GridBagConstraints gbc = new GridBagConstraints();
-        gbc.insets = new Insets(10, 10, 10, 10); 
+        gbc.insets = new Insets(10, 10, 10, 10); // Spacing between fields
         gbc.fill = GridBagConstraints.HORIZONTAL;
 
         JLabel l1 = new JLabel("Username:");
@@ -89,47 +82,55 @@ public class addIns {
         gbc.anchor = GridBagConstraints.WEST;
         p2.add(tPassword, gbc);
 
-        JLabel l3 = new JLabel("Department:");
+        JLabel l3 = new JLabel("Program:");
         l3.setFont(gFont.deriveFont(Font.PLAIN, 24));
         l3.setForeground(Color.decode("#020A48"));
         gbc.gridx = 0; gbc.gridy = 2; gbc.weightx = 0;
         gbc.anchor = GridBagConstraints.EAST;
         p2.add(l3, gbc);
 
-        JTextField tDepartment = new JTextField(20);
-        tDepartment.setFont(gFont.deriveFont(Font.PLAIN, 21));
+        JTextField tProgram = new JTextField(20);
+        tProgram.setFont(gFont.deriveFont(Font.PLAIN, 21));
         gbc.gridx = 1; gbc.gridy = 2; gbc.weightx = 1;
         gbc.anchor = GridBagConstraints.WEST;
-        p2.add(tDepartment, gbc);
+        p2.add(tProgram, gbc);
 
-        // --- Empty Row 3 for spacing (to push buttons down) ---
-        // This is done implicitly by adjusting the button panel inset, but adding a spacer helps center the form better
-        
-        // --- Row 4: Buttons ---
+        JLabel l4 = new JLabel("Year:");
+        l4.setFont(gFont.deriveFont(Font.PLAIN, 24));
+        l4.setForeground(Color.decode("#020A48"));
+        gbc.gridx = 0; gbc.gridy = 3; gbc.weightx = 0;
+        gbc.anchor = GridBagConstraints.EAST;
+        p2.add(l4, gbc);
+
+        JTextField tYear = new JTextField(20);
+        tYear.setFont(gFont.deriveFont(Font.PLAIN, 21));
+        gbc.gridx = 1; gbc.gridy = 3; gbc.weightx = 1;
+        gbc.anchor = GridBagConstraints.WEST;
+        p2.add(tYear, gbc);
+
+        // --- LOWS ---
         JPanel buttonPanel = new JPanel(new FlowLayout(FlowLayout.CENTER, 20, 0));
         buttonPanel.setOpaque(false);
- 
 
-        JButton b1 = new JButton("Back");
-        b1.setBackground(Color.decode("#2f77b1")); 
-        b1.setForeground(Color.WHITE); 
-        b1.setFont(breatheFont.deriveFont(Font.PLAIN, 35));
-        b1.setMargin(new Insets(10, 30, 5, 30));
-        buttonPanel.add(b1);
+        JButton bBack = new JButton("Back");
+        bBack.setBackground(Color.decode("#2f77b1")); 
+        bBack.setForeground(Color.WHITE); 
+        bBack.setFont(breatheFont.deriveFont(Font.PLAIN, 35));
+        bBack.setMargin(new Insets(10, 30, 5, 30));
+        buttonPanel.add(bBack);
 
-        JButton b2 = new JButton("Register");
-        b2.setBackground(Color.decode("#2f77b1")); 
-        b2.setForeground(Color.WHITE); 
-        b2.setFont(breatheFont.deriveFont(Font.PLAIN, 35));
-        b2.setMargin(new Insets(10, 30, 5, 30));
-        buttonPanel.add(b2);
+        JButton bRegister = new JButton("Register");
+        bRegister.setBackground(Color.decode("#2f77b1")); 
+        bRegister.setForeground(Color.WHITE); 
+        bRegister.setFont(breatheFont.deriveFont(Font.PLAIN, 35));
+        bRegister.setMargin(new Insets(10, 30, 5, 30));
+        buttonPanel.add(bRegister);
 
-        // Added to row 3 (since we removed row 3/4 from addStd)
-        gbc.gridx = 0; gbc.gridy = 3; 
+        gbc.gridx = 0; gbc.gridy = 4; 
         gbc.gridwidth = 2; 
         gbc.anchor = GridBagConstraints.CENTER;
         gbc.fill = GridBagConstraints.NONE;
-        gbc.insets = new Insets(120, 10, 10, 10); 
+        gbc.insets = new Insets(60, 10, 10, 10); 
         p2.add(buttonPanel, gbc);
 
         f.add(p2, BorderLayout.CENTER);
@@ -139,20 +140,22 @@ public class addIns {
         f.setVisible(true);
 
         // ---- Action Listeners ----
-        b2.addActionListener(new ActionListener() {
+        bRegister.addActionListener(new ActionListener() {
             public void actionPerformed(ActionEvent e) {
-                if (tUsername.getText().isEmpty() || tPassword.getText().trim().isEmpty() || tDepartment.getText().trim().isEmpty()) {
+                if (tUsername.getText().isEmpty() || tPassword.getText().trim().isEmpty() || 
+                    tProgram.getText().trim().isEmpty() || tYear.getText().trim().isEmpty()) {
                     JOptionPane.showMessageDialog(null, "All fields must be filled out.", "Error", JOptionPane.ERROR_MESSAGE);
                     return;
                 }
 
                 String username = tUsername.getText();
-                String department = tDepartment.getText().trim(); 
-                String hash_pass = HashGenerator.makeHash(tPassword.getText().trim());
+                String hashPass = HashGenerator.makeHash(tPassword.getText().trim()); 
+                String program = tProgram.getText().trim();     
+                String year = tYear.getText().trim();                        
 
-                AdminService service = new AdminService();
-                String result = service.registerInstructor(username, hash_pass, department);
-                
+                AdminService service = new AdminService();  
+                String result = service.registerStudent(username, hashPass, program, year);
+
                 if (result.startsWith("Success")) {
                     JOptionPane.showMessageDialog(null, result, "Success", JOptionPane.INFORMATION_MESSAGE);
                     f.dispose();
@@ -165,12 +168,12 @@ public class addIns {
             }
         });
 
-        b1.addActionListener(new ActionListener() {
+        bBack.addActionListener(new ActionListener() {
             public void actionPerformed(ActionEvent e) {
-                System.out.println("\tGoing back to Admin Dashboard..");
+                System.out.println(" to Admin Dashboard..");
                 new adminDashboard(rollNo);
                 f.dispose();
             }
-        });    
+        });
     }
 }
