+++ src/edu/univ/erp/util/transcript_7.csv
@@ -1,0 +1,3 @@
+"akshainie Result for semester THE FALL and year 2025"
+"Course Code","Section","Title","Credits","Grade"
+"IP1111","B","Instroduction to plumbing","4","B"
